// src/synapse/types.ts
export interface MasPaginationLinks {
  self?: string;
  first?: string;
  last?: string;
  next?: string;
  prev?: string;
}

export interface MasPageMeta {
  page?: {
    cursor?: string;
  };
}
// ...
export interface MASRegistrationTokenAttributes {
  token: string;
  valid: boolean;
  usage_limit?: number;
  times_used: number;
  created_at: string;
  last_used_at?: string;
  expires_at?: string;
  revoked_at?: string;
}

export interface MASRegistrationTokenResource {
  type: string;
  id: string;
  attributes: MASRegistrationTokenAttributes;
  meta?: MasPageMeta;
  links: {
    self: string;
  };
}

export interface MASRegistrationToken {
  data: MASRegistrationTokenResource;
  links: {
    self: string;
  };
}

export interface MASRegistrationTokenListResponse {
  data: MASRegistrationTokenResource[];
  meta?: {
    count?: number;
  };
  links?: MasPaginationLinks;
}